Fix all compiler warnings in the ESP-IDF build
The component emitted 21 warnings, all format mismatches.

-Wformat= (19): on ESP32 uint32_t is `long unsigned int`, so %u/%d did
not match the millisecond timers, timestamps, start counts, retry delay
and BLE passkey being logged. Switched to PRIu32 from <cinttypes>.
ESPTime::timestamp is now printed via a `long long` cast, since time_t
is `long long` on ESP-IDF but `long` on many hosts and the sources are
also compiled on the host by the test suite.

-Wformat-truncation= (2): the "HH:MM" helpers in schedule_entry.h write
into char[6], and the compiler cannot prove a uint8_t hour/minute
renders in two digits. Bounded both fields so the buffer can never
truncate, which also keeps the output well-formed if the pump ever
reports an out-of-range byte.

No behavioural change: only format specifiers and the values passed to
them changed.

Tested:
- Full ESP-IDF build (ESPHome 2026.7.0, esp32-c3): 0 warnings, was 21.
  All previously-warning translation units confirmed recompiled.
- Host suite `cd tests && make clean && make test` (-Wall -Wextra):
  0 warnings, 494 assertions passing, 0 failures.

+++ components/alpha_hwr/time_service.cpp
@@ -210,8 +210,10 @@ ESPTime TimeService::parse_clock_response(const uint8_t *data, size_t len) {
   base.recalc_timestamp_local();
   pump_time.timestamp = base.timestamp;
   
-  ESP_LOGD(TAG, "ESPTime created: timestamp=%ld, is_valid=%d", 
-           pump_time.timestamp, pump_time.is_valid());
+  // time_t width varies by platform (long long on ESP-IDF, long on many hosts):
+  // widen to long long so one format string is correct everywhere.
+  ESP_LOGD(TAG, "ESPTime created: timestamp=%lld, is_valid=%d",
+           static_cast<long long>(pump_time.timestamp), pump_time.is_valid());
   
   return pump_time;
 }
